Fixes git-ftp's host argument.
According to docs at https://github.com/git-ftp/git-ftp, it was
necessary to prefix the FTP url with "ftp://" and also either append the
remote root folder in the FTP URL or provide the --remote-root argument.

If not specified, the script will use the FTP protocol. If the protocol
is provided with the ftpHost string, it will follow the protocol name as
long as they are one of the supported protocols in git-ftp:
- ftp
- sftp
- ftps
- ftpes

Notice the 'sftp' protocol might require SSH keys or known hosts
file/hashes; this script currently does not support providing these
settings.