Restructures the main job to send slack notifications earlier.
It will try to assemble Slack notification first so that it could send
slack notifications about issues with all non-slack-specific
configuration.

The parameters handling code was then moved within the main async loop
so that it can call the asynchronous slack messaging methods.

An special configError() method was implemented to craft the slack
notifications as it fails the script execution on errors.