Adds back initialize feature if push fails -- with changes.
Restores the "git-ftp init" command call if the push fails.

The change is that now it will be called if git-ftp fails for not
finding its control file (.git-ftp.log), instead for if the remote
directory does not exist. It shall still fail if the remote directory is
not present.